Extract NilaiSikapRequest into its own file
